smart-speaker-gemini/pyproject.toml
[project]
name = "smart-speaker-gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "pyaudio>=0.2.14",
    "pygame>=2.6.1",
    "python-dotenv>=1.1.0",
    "pyttsx3>=2.98",
    "requests>=2.32.3",
    "speechrecognition>=3.14.3",
]
